+++ src/main/java/pl/plajer/villagedefense/handlers/party/PartiesPartyHandlerImpl.java
@@ -22,11 +22,10 @@ import com.alessiodp.parties.api.Parties;
 import com.alessiodp.parties.api.interfaces.PartiesAPI;
 import com.alessiodp.parties.api.interfaces.Party;
 import com.alessiodp.parties.api.interfaces.PartyPlayer;
-
-import java.util.stream.Collectors;
-
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
+
+import java.util.stream.Collectors;
 
 /**
  * @author Plajer
@@ -38,7 +37,11 @@ public class PartiesPartyHandlerImpl implements PartyHandler {
   @Override
   public boolean isPlayerInParty(Player player) {
     PartiesAPI api = Parties.getApi();
-    return api.getPartyPlayer(player.getUniqueId()) != null;
+    PartyPlayer partyPlayer = api.getPartyPlayer(player.getUniqueId());
+    if (partyPlayer == null) return false;
+    Party party = api.getParty(partyPlayer.getPartyName());
+    if (party == null) return false;
+    return party.getMembers().size() > 1;
   }
 
   @Override