+++ CHANGELOG.md
@@ -11,6 +11,9 @@ That's all, matcher will stop when detects next line started with `###` match
 
 ### 4.4.1 Release 
 * Fixed leave item events not working
+* Added missing string to default bungee.yml
+* Fixed new messages from last update
+* Added back all messages to language.yml (Upgrades and Holograms)
 
 ### 4.4.0 Release (23.01.2020 - 18.03.2020)
 * Added new Zombie VillagerBuster (wave 15 and up)