+++ CHANGELOG.md
@@ -13,6 +13,7 @@ That's all, matcher will stop when detects next line started with `###` match
 * Fixed leave item events not working
 * Added missing string to default bungee.yml
 * Fixed new messages from last update
+* Tried to fix parties
 * Added back all messages to language.yml (Upgrades and Holograms)
 
 ### 4.4.0 Release (23.01.2020 - 18.03.2020)